+++ examples/fcu_motor_test.cpp
@@ -16,7 +16,7 @@
 
 int main(int argc, char *argv[]) {
     const std::string device = (argc>1) ? std::string(argv[1]) : "/dev/ttyUSB0";
-    const uint baudrate = (argc>2) ? std::stoul(argv[2]) : 115200;
+    const size_t baudrate = (argc>2) ? std::stoul(argv[2]) : 115200;
 
     fcu::FlightController fcu(device, baudrate);
     fcu.initialise();
